// types/enums.ts
export enum BlogStatus {
  PUBLISHED = "published",
  PENDING = "pending",
}

export enum Categories {
  RUBY_ON_RAILS = "Ruby on Rails",
  REACT = "React",
  NEXTJS = "Next.js",
  FRONTEND = "Frontend",
  BACKEND = "Backend",
  GITHUB = "GitHub",
  CI_CD = "CI/CD",
  FULLSTACK = "Fullstack",
  DESIGN_PATTERNS = "Design Patterns",
  TESTING = "Testing",
}